Updated the Timer class
The update was required for more reasons. On the one hand the functionality had to be prepared so a derived class could be efficiently created for a different timer type (e.g. repeating timer). Some additional functions were required for this (e.g. setter for timeout period) and a few existing ones had to be virtualized (start, restart, stop). On the other hand some functionality had to be slightly changed, i.e. the provided timeout period had to persist after starting or restarting the timer