Moved the main() code into a macro
This was required for easily reusing the code for testing other tree types (e.g. red-black trees) without the need to refactor the main() code into functions. The main() code is just a pure-testing code and there is no point for the moment in refactoring it into a more elegant solution