Refactored the removal function for red-black tree
It was required to replace the calls to rotation functions of class (RedBlack)Node with calls to rotation functions of class RedBlackTree/BinarySearchTree. The reason is that the former ones don't check if the root node is rotated which might lead to bugs/crashes. Also it was required to rename a boolean variable as the old name was misleading (recurring occured even when this variable was not used). Finally it was needed to reduce the nesting level of if-else instructions