Updated the move assignment operator for simple BST objects
It was required to update this operator as the simple BST objects might "impersonate" derived objects by using BinarySearchTree pointers to derived BSTs. Thus it would be possible to move assign different object types. This might cause errors and inconsistencies if the move assignment was kept in its original implementation. To ensure correctness it is necessary to first copy assign the values of the source tree and then erase the source tree content

Also it was required to create a move assignment test for mixed tree types to demonstrate that this update of the move assignment operator is successful

As a final note the move assignment operator of derived tree objects does not need to be updated as it is not possible to use it for doing move assignment operations between different tree type between different tree types. Also the RedBlackTree and AVLTree classes are final