Updated the settings of the unit tests
It was required to create an init() and cleanup() function for each BST test case. Also it was required to update the constructor and destructor and to rename the _reset() function in order to point out what this method is effectively supposed to do. The destructor functionality had to be taken over by the cleanup() function