Refactored the Kruskal and Prim tree algorithms
It was required to consolidate the algorithms into a single project (with 2 executables) in order to minimize the amount of duplicate code. Also it was required to solve a crash within Prim algorithm by taking loose tree parts into consideration (e.g. a node that is not connected to the graph)