Implemented simple BST with smart pointers
This was required in order to better understand how the smart pointers
(shared_ptr, weak_ptr) work and to enhance the understanding of the
ownership concept

Currently only the simple binary search trees have been updated. The
reason is that for red-black and AVL trees there are some covariant
virtual functions that could not be resolved with smart pointers.