Updated DataOrdering IO functionality
It was required to create a structured input/output approach by using a reader and a writer class that are managing the IO resources. Also it was required to refactor the IO methods by using overloaded stream operators in order to achieve more manageable code. Last but not least it was needed to move the ordered set calculation to DataOrderingEngine in order to obtain a better separation of concerns/responsabilities